feat(items): update items api

+++ src/api/items.ts
@@ -1,21 +1,24 @@
 import type { NextFunction, Request, Response } from 'express'
-import { constants } from 'fs'
-import fs from 'fs/promises'
-import { deleteImage, uploadImage } from '../cloudinary.js'
+import { access, constants, rm } from 'fs/promises'
+import { isValidObjectId } from 'mongoose'
+import { deleteFromCloudinary } from '../cloudinary.js'
 import { AppError, Errors } from '../error.js'
 import { ItemModel } from '../models/item.js'
-import { redis } from '../redis.js'
 
-export async function getItems(req: Request, res: Response, next: NextFunction) {
-  const items = await ItemModel.find({ owner: req.auth?.data._id }, '-owner').lean()
+export async function getItems(req: Request, res: Response, _next: NextFunction) {
+  const items = await ItemModel.find({ owner: req.auth?.data._id }).lean()
   res.send(items)
 }
 
 export async function getItemById(req: Request, res: Response, next: NextFunction) {
-  const item = await ItemModel.findOne({ _id: req.params._id, owner: req.auth?.data._id }, '-owner').lean()
-  console.log(item)
+  const { _id } = req.params
+  if (!isValidObjectId(_id)) {
+    return next(new AppError(Errors.ITEMS.COMMON.INVALID_ID))
+  }
+
+  const item = await ItemModel.findOne({ _id, owner: req.auth?.data._id }, '-owner').lean()
   if (!item) {
-    return next(new AppError(Errors.ITEMS.GET_BY_ID.ITEM_NOT_FOUND))
+    return next(new AppError(Errors.ITEMS.COMMON.ITEM_NOT_FOUND))
   }
 
   res.send(item)
@@ -29,39 +32,22 @@ export async function postItem(req: Request, res: Response, next: NextFunction) 
     return next(new AppError(Errors.ITEMS.POST_ITEM.LIMIT_REACHED))
   }
 
-  let uploadedImages = []
-
-  // upload images to Cloudinary
-  try {
-    uploadedImages = await uploadImage(images)
-  } catch (error) {
-    return next(new AppError(Errors.ITEMS.POST_ITEM.UPLOAD_IMAGES, error))
-  }
-
   const newItem = new ItemModel({
     name,
     icons,
-    images: uploadedImages.map((image) => image.url),
+    images,
     tags,
     owner: req.auth?.data._id,
   })
 
+  await newItem.validate()
+
   try {
     const item = await newItem.save()
-    // delete images from disk
-    images.map((image: string) => {
-      redis.del(image)
-      fs.access(`${global.__basedir}${image}`, constants.F_OK)
-        .then(() => fs.rm(`${global.__basedir}${image}`))
-        .catch(() => `Failed to delete ${global.__basedir}${image}, file not found`)
-      fs.access(`${global.__basedir}${image}.webp`, constants.F_OK)
-        .then(() => fs.rm(`${global.__basedir}${image}.webp`))
-        .catch(() => `Failed to delete ${global.__basedir}${image}.webp, file not found`)
-    })
 
     return res.send(item)
-  } catch (error: any) {
-    return next(new AppError(Errors.ITEMS.POST_ITEM.ITEM_VALIDATION_ERROR, error.message))
+  } catch (error: unknown) {
+    return next(new AppError(Errors.ITEMS.POST_ITEM.ITEM_VALIDATION_ERROR, error))
   }
 }
 
@@ -69,28 +55,52 @@ export async function editItem(req: Request, res: Response, next: NextFunction) 
   const { _id } = req.params
   const { name, icons, tags } = req.body
 
-  const updatedItem = await ItemModel.findOneAndUpdate({ _id, name, owner: req.auth?.data._id }, { icons, tags })
+  if (!isValidObjectId(_id)) {
+    return next(new AppError(Errors.ITEMS.COMMON.INVALID_ID))
+  }
+
+  const updatedItem = await ItemModel.findOneAndUpdate({ _id, owner: req.auth?.data._id }, { name, icons, tags })
+  if (!updatedItem) {
+    return next(new AppError(Errors.ITEMS.COMMON.ITEM_NOT_FOUND))
+  }
 
   return res.send(updatedItem)
 }
 
 export async function deleteItem(req: Request, res: Response, next: NextFunction) {
   const { _id } = req.params
+  if (!isValidObjectId(_id)) {
+    return next(new AppError(Errors.ITEMS.COMMON.INVALID_ID))
+  }
 
   const deletedItem = await ItemModel.findOneAndRemove({ _id, owner: req.auth?.data._id })
-  if (deletedItem) {
-    try {
-      const ids = deletedItem.images.map((image) => {
-        const imageName = image.split('/').pop()
-        if (!imageName) return ''
-        const imageId = imageName.split('.')[0]
-        return imageId
-      })
+  if (!deletedItem) {
+    return next(new AppError(Errors.ITEMS.COMMON.ITEM_NOT_FOUND))
+  }
 
-      deleteImage(ids)
+  const isCloudImages = deletedItem.images.some((image) => image.indexOf('cloudinary') !== -1)
+
+  if (isCloudImages) {
+    const idsForDelete: Array<string> = []
+    deletedItem.images.forEach((url) => {
+      const pathname = new URL(url).pathname
+      const image = pathname.split('/').pop()
+      const id = image?.split('.')[0]
+
+      if (id) {
+        idsForDelete.push(id)
+      }
+    })
+
+    try {
+      deleteFromCloudinary(idsForDelete)
     } catch (error) {
       console.error(error)
     }
+  } else {
+    deletedItem.images.forEach((path) => {
+      access(`${global.__basedir}${path}`, constants.F_OK).then(() => rm(`${global.__basedir}${path}`))
+    })
   }
 
   return res.send(true)
